[fix] #33 - 친구 관리 테이블 뷰 셀 disclosureIndicator 추가 코드 수정
